fix: 중복 로직 커스텀 훅 추출 및 라벨 필터 스타일 수정 (#48)
