+++ src/components/thread-header.tsx
@@ -1,8 +1,5 @@
-import { useMemo } from "react"
-
 import { Badge } from "@/components/ui/badge"
 import { AccountIcon } from "@/lib/icon-entries"
-import { useLabels } from "@/queries/labels"
 import type { LabelSummary } from "@/types/email"
 import type { MailAccount } from "@/types/mail-account"
 
@@ -21,9 +18,6 @@ export function ThreadHeader({ thread, account, labels }: ThreadHeaderProps) {
   const messageCount = thread.messages.length
   const hasInbound = thread.messages.some((m) => m.direction === "INBOUND")
   const hasOutbound = thread.messages.some((m) => m.direction === "OUTBOUND")
-  const { data: labelsList } = useLabels()
-  const labelsColorMap = useMemo(() => new Map(labelsList?.map((l) => [l.id, l.colorCode]) ?? []), [labelsList])
-
   return (
     <div className="shrink-0 border-b px-6 pt-2 pb-5">
       <h2 className="text-xl leading-snug font-semibold wrap-break-word">{thread.latestSubject || "(제목 없음)"}</h2>
@@ -55,7 +49,7 @@ export function ThreadHeader({ thread, account, labels }: ThreadHeaderProps) {
           <span
             key={label.labelId}
             className="inline-flex items-center rounded-full px-1.5 py-0.5 text-xs font-medium text-white"
-            style={{ backgroundColor: labelsColorMap.get(label.labelId) ?? label.colorCode }}
+            style={{ backgroundColor: label.colorCode }}
           >
             {label.name}
           </span>
